Setting up the auth on routes.

+++ server/src/routes/memories.ts
@@ -3,8 +3,17 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  // the JWT verification will occur before each one of these routes
+  app.addHook("preHandler", async (request) => {
+    // verifies if the token is present in the current request
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub, // user's id
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -19,7 +28,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, response) => {
     // this zod schema is responsible to validate
     // if the id param from the request object is a string
     const paramsSchema = z.object({
@@ -34,6 +43,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
 
     return memory;
   });
@@ -54,17 +67,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "e42f8314-b5e6-4979-9844-e373beecdf02",
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, response) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
+
+    const { id } = paramsSchema.parse(request.params);
 
     const bodySchema = z.object({
       content: z.string(),
@@ -72,10 +87,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     });
 
-    const { id } = paramsSchema.parse(request.params);
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -89,12 +113,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, response) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
 
     // delete the memory that has the id param
     await prisma.memory.delete({
